+++ Source/Entities/SoundContainer.cpp
@@ -282,6 +282,11 @@ void SoundContainer::SetPosition(const Vector& newPosition) {
 	}
 }
 
+float SoundContainer::GetAudibleVolume() const {
+	return g_AudioMan.GetSoundContainerAudibleVolume(this);
+}
+
+
 void SoundContainer::SetVolume(float newVolume) {
 	newVolume = std::clamp(newVolume, 0.0F, 10.0F);
 	if (IsBeingPlayed()) {